WIP: Move parent checks to onStatementAdded()
We are performing parent structural checks during buildEffective(),
which can be performed as soon as the statement is added.

The notable problem is checking for list with a key: while the list
is our parent, and therefore obviously exists at that time, key
statement's order may be such that it is encountered only afterwards.

This can be solved rather easily by placing an inference action
to request the list's full definition, at which point all of its
substatements are guaranteed to be present.

// FIXME: actually do the above

JIRA: YANGTOOLS-1186
Change-Id: Id3e8cf6477765c3f89f657fd770952a7035ec876
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>