WIP: Make BuildGlobalContext thread-local
Acquiring a namespace behavior from ReactorStmtCtx goes through
a round of indirections through RootStatementContext. These are
accessed quite frequently and relying on getRoot() as it involves
walking along parent axis.

Create a thread-local variable which holds the currently-executing
BuildGlobalContext and route invariants. This makes it obvious that
SourceSpecificContext actually is not a NamespaceBehaviour.Registry,
cutting one implementation down.

// FIXME: measure performance impact

JIRA: YANGTOOLS-1204
Change-Id: I8f8df58d575039152f47fc48cbd6c2687cab4f69
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>