WIP: Add NamespaceStorageNode.getGlobalNamespaceStorage()
We have a number of global namespaces, which are stored in
BuildGlobalContext only. Various levels of the statement reactor
tree need to access these through walk along parent storage
axis.

Since we are accessing these frequently and we already have
a thread-local BuildGlobalContext, add a shortcut accessor and
take advantage of it in NamespaceStorageSupport.Global.

// FIXME: measure impact

JIRA: YANGTOOLS-1159
Change-Id: I74ac1a31176e8f3fc33875dff775965d1129391f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>