Cache substatement context-independence summary
Individual statements expose their statically-known CopyPolicy,
which governs how they react to being copied. This does not account
the statement instance's substatements, as that is a
partially-dynamic property.

Even if a particular statement is affected by the copy operation,
it is quite common for its substatement (tree) to not be affected --
lending itself for reuse of the effective substatement list as well
as any structures which are derived from it.

Whenever a statement completes ModelProcessingPhase.EFFECTIVE_MODEL,
examine all substatements and whether all of them are recursively
context-independent.

Expend our last remaining bit in ReactorStmtCtx.flags to remember
the check's result -- preventing the need to recurse into a
substatement's substatements.

JIRA: YANGTOOLS-1209
Change-Id: I0fda715da441038a1276678f49718df9bfacfb90
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>