Reuse AbstractTypeStatementSupport's statements
This looks like a very trivial conversion. It looks very much we
have a ton of dead code here, as the copy mechanics does not
seem to trigger at all.

At any rate, we have we need for completess' sake. We may end up
dropping it at a later time.

JIRA: YANGTOOLS-1213
Change-Id: I7aa6ce794def1e19a8adf6bc7300303f1bfdde65
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>