YangStmtMapping.TYPE implies StatementPolicy.contextIndependent()
InferredStatementContext has an explicit guard to get replicas of
statements which are YangStmtMapping.TYPE. This amounts to those
statements being context-independent. Express that fact in the
policy.

This exposes an issue with 'base' statements, which in fact can
be considered for copying and are really context-indepenent, not
uncopyable.

JIRA: YANGTOOLS-1208
Change-Id: I49f0454ad11a029d217bd7fb70d23e315f37072c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>