Fix uses-refine statement reference counting
When a statement is being added as a substatement, it is the caller's
responsibility to ensure proper refcount is kept.

UsesStatementSupport.addOrReplaceNode() failed to create a replica,
which resulted in an InferredStatementContext containing a
SubstatementContext -- which logically should never happen.

This issue has been there for quite some time, but was masked by
eager context instantiation. Fix it by acquiring a replicaAsChildOf(),
just as other callers of addEffectiveSubstatement() do.

JIRA: YANGTOOLS-1208
Change-Id: I88aa1009a64a4fb6df9138f748877836bfddd175
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>