ONE: yang-data validation is an inference action
We are accessing inference state during effective build, which is
a no-no, with all sorts of potential rabbits -- and we have
a continent which can tell a tale or two about that.

So realize that we just need to perform a simple substatement check
and be done with it.

This implementationallows other data nodes being there -- which is
in keeping with the spec. If somebody complains, the flip side
of yang-data is that it is only its sole container statement which
is accessed. In other words users are fully expected to ignore
other statements and treat that container as schema tree root.

JIRA: YANGTOOLS-1186
Change-Id: I802193d75d914216ce565f51c042547abb97caba
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>