Do not build unrecognized extension statements
Attempting to build EffectiveStatements for extensions we do not
recognize is an exercise in futility: we have to make assumptions
about arguments (like whether they can form a QName), about copy
policy (how does the extension propagate through uses?).

We also need to deal with extensions being used recursively, for
which we use a mighty object leaking machinery.

This patch changes the policy by explicitly taking out
UnrecognizedEffectiveStatement as something we normally produce
and making sure that if we do not have a StatementSupport registered
for a statement, it will not propagate into the Effective world.

Any extensions encountered are still retained in DeclaredStatement
form, as we can understand that definition well enough.

JIRA: YANGTOOLS-1196
Change-Id: If4e2292a977f8be82de7dae98671f025b79a836e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>