Add SchemaInfereceStack.enterDataTree()
SchemaInferenceStack is a very useful tool in figuring things out
around EffectiveStatement tree. So far it supported only simple
entries along 'grouping' and 'schema tree' namespaces.

Add support for walking the 'data tree' namespace as well, which
in itself is trivial, but has major implications for the conversion
paths to both SchemaPath and SchemaNodeIdentifier.

JIRA: YANGTOOLS-1231
Change-Id: I8ea288ede962000d54cf8fac8863f8bc7bd3be7f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>