Add the ability to retain DeclarationReference
There are use cases where we would like to retain information about
where a statement was originally declared -- for example for use in
IDEs.

We already have a parser-internal representation of most of this
information in the form of StatementSourceReference, but that is not
retained in DeclaredStatements nor is it a public API.

Introduce DeclarationReference and its two specializations,
DeclarationInText and DeclarationInFile. These can be acquired via
DeclaredStatement.declarationReference().

Also provide integration between DeclarationReference and
StatementSourceReference, which will allow yang-parser-reactor to turn a
source reference into a declaration reference.

Finally add the basic substrate needed for forwarding decorators
that attach DeclarationReference to a delegate.

JIRA: YANGTOOLS-1193
Change-Id: I4e4d9567c83c08aab9de2a2685470835fadcd0d7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>