Define ExecutionOrder
ModelProcessingPhase really amounts to an execution order, in that
the phases need to be executed serially.

Let's define a space for 127 values of ModelProcessingPhase by
defining ExecutionOrder, which is semantically the same thing (but
different).

Update StatementContextBase to store the execution order as single
byte instead of a reference. Doing so allows us to simplify some
of the verification logic, but most importantly it frees up
some space -- resulting in 3-7 bytes worth of internal space losses.

JIRA: YANGTOOLS-1254
JIRA: YANGTOOLS-1150
Change-Id: I3e5b23cae45798b4508dd9e270a4b0317613ed1b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>