Update module-info.java
Do not rely on transitive access to packages, spell them out. This makes
the development experience with Eclipse a bit better, as it does not end
up adding random requires.

Also hide a number of private constructs, so we do not need to re-export
imports where not needed.

JIRA: YANGTOOLS-932
Change-Id: I32431065724dee226754bfbd60ca70b839c4f49f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>