TypeStatement takes a QName argument
YANG 'type' statement has effective semantics attached to the
namespace where it is defined. These need to be captured into a
correct QName, so that users trying to infere on top of
TypeEffectiveStatement do not need to deal with interpreting
the value.

We even make it exceedingly difficult, as the statements have an
'exact replica' policy, hence the argument namespace is no way
related to parent namespace.

Change argument mapping of 'type' from String to QName. Introduce
the corresponding EffectiveStatementNamespace as well.

This changes the behavior of nested union typedefinitions, where
we would be confused as to the nested base type's namespace: since
it is a plain reference, it has to point to the YANG namespace,
not the model's.

JIRA: YANGTOOLS-1117
Change-Id: Ib44cb2796ba11de4a20d86c382d25e753c61279e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>