DataContainerNode is not generic
DataContainerNode and its specializations are using generic type
to capture the result of getIdentifier(). This is not necessary,
as we can achieve the same effect by properly overriding
getIdentifier().

This leads to a simpler object model and better ergonomics, as
a significant number of wildcards are removed from related interfaces,
such as NormalizedNodeContainer and DistinctNodeContainer.

JIRA: YANGTOOLS-1022
Change-Id: Ibafca25dcdc20911ed5048776d1b7a6429ccf3cf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>