Add SchemaInferenceStack.enterSchemaTree(SchemaNodeIdentifier)
SchemaNodeIdentifier is a well-known concept which we are integrating
with in multiple aspects. One key component is the ability to resolve
a SchemaNodeIdentifier in the context of an existing stack. While users
can do this easily themselves, it is more natural to provide this
ability out of the box.

JIRA: YANGTOOLS-1231
Change-Id: I9818f175d09fd4e6e4f610127929ee8553c3116e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>