Cleanup unrecognized statement wrapping
reactor/rfc7950 interaction around inner statements is a bit hacky
in that we end up performing two unrelated semantic checks before
dispatching.

As it turns out, though, this only a special case of a more widely
applicable pattern, where the parent statement chooses to wrap its
children in a different semantic support.

Separating this out makes the intent clear, as it ends up being an
extension trait to StatementSupport on the side of yang-parser-reactor
and has a well-understood anchor in UnrecognizedStatementSupport.

Change-Id: I3b64f58239e22affccc02921efd8cd70b411d7c1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>