Teach YANG parser to retain DeclarationReference
DeclaredStatement is exposing DeclarationReference as optional
additional information. yang-parser-reactor has knowledge of this
information during its execution to aid diagnosing model errors, so we
just need a mechanism for to bridge this information to statement
implementations.

Add the end-to-end glue code for yang-parser-{api,reactor} users
to  request this information be retained and that intent to be
communicated to individual StatementFactories. Each StatementFactory
needs to opt-in by providing the appropriate decorator via the substrate
provided in yang-model-spi's AbstractRefStatement.

JIRA: YANGTOOLS-1193
Change-Id: I6a10feb682bb720349e927ea171593e663278eea
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>