Update yang-parser-api
Current API is hosted in yang.model.parser.api package, which is a
mouthful, with 'model' being the redundant component, really.

Introduce a new set of APIs in yang.parser.api package, improving
documentation as well as clarifying a few concepts. Since
YangParserConfiguration is a new concept, we move it directly to its new
home.

JIRA: YANGTOOLS-1274
Change-Id: I2149c9d88a7724059b4cc5c56fbc6c192164c99a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>