Fixes the missing augmented mandatory leaf in choice
MandatoryLeafEnforcer fails to find mandatory leaf which
was added by augmentation. It utilizes NormalizedNodes.findNode()
when looking for mandatory leafs. NormalizedNodes looks through
direct children and ignores the case, where the child is and
instance of AugmentationNode encapsulating other leafs.

This patch allows the NormalizedNodes.getDirectChild() to look
for the queried node in all AugmentationNode children.

Also adding minimized model to showcase the problematic
augmentation and 2 tests to showcase:
- writing data with the augmented nodes as direct children
- writing the same data in an AugmentationNode

Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>
Change-Id: I3ea81e5affa1b6a8a187179c4500838a247ec95f