Restore special-case handling of NotificationDefinition
We have historically had a weird special case for parsing notifications,
where we would expect the stack to be initialized to point to the
notification and then to enter it as an alternative to normal data tree.

This functionality is used in NETCONF, restore it back in a slightly
better format than before. Also drop a FIXME so we expose the
functionality in a more structure way in the future.

JIRA: YANGTOOLS-1288
Change-Id: I9237cf558525e4028626cfe7f67d76bbabebab4c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>