Add DistinctNodeContainer.getChildByArg()
There are a number of places where there is an implied assertion
that the child must exist, but the caller does not care about the
exception being reported.  These are using
findChildByArg().orElseThrows() today, could use a more expressive
alternative.

Add getChildByArg() as a default method, which wraps a call to
childByArg() with a verifyNotNull(), neatly expressing the intent.

Change-Id: I8dade860a0c615e987aa2c017473eb08769afee3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>