Fix anyxml-to-JSON translation duplication error
JSONNormalizedNodeStreamWriter generates duplicates when there are array
elements are not listed consecutively and interleaved with other sibling
nodes. Make sure to check if there is any previously processed sibling
node with the same node name instead of checking just the previous one.

Also, Add a couple of test cases to cover this issue.

JIRA: YANGTOOLS-1302
Change-Id: I4d64c8b8a45009e7667af621c96c34824fb7b48b
Signed-off-by: Manoj Chokka <cmanoj8@gmail.com>
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>