Correct DefaultWith{Schema,Data}Tree statefulness
We have a slight thinko here: we retain substatements in
WithSubstatements subclass, hence the base class really says it has
empty substatements, which is implying both schema tree and data tree
are empty.

Since our baseline is to return an empty map for namespace contents,
this is completely unneeded. Eliminate the difference by eliminating the
WithSubstatements specialization.

Change-Id: I35a7f16910bbda7c14faf0b9b69fabc84ed06f9a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>