Correct DefaultWith{Schema,Data}Tree statefulness
We have a slight thinko here: we retain substatements in
WithSubstatements subclass, hence the base class really says it has
empty substatements, which is implying both schema tree and data tree
are empty.

Move the three maps to WithSubstatements, which allows us to better
specialize empty containers, if that need ever arises.

// FIXME: further reduce classes
// FIXME: create an issue for this
// FIXME: refactor only after moving leaf-list statements

Change-Id: I35a7f16910bbda7c14faf0b9b69fabc84ed06f9a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>