Promote UnknownEffectiveStatementBase to yang-model-spi
The semantics of UnknownSchemaNode has been cleaned up and documented a
bit more. We can therefore update UnknownEffectiveStatementBase and
promote it to yang-model-spi, as we no longer need dependencies on
yang-parser-spi. Also rename it to AbstractEffectiveUnknownSchemaNode to
make to ensure we have proper naming. Since this is the only subclass of
AbstractEffectiveDocumentedNodeWithStatus, we are combining the two,
further simplifying things.

JIRA: YANGTOOLS-1320
Change-Id: I090437cc5958ef1f9bacd80d899f2a3c08dd8ded
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>