Introduce UnresolvedQName
Our AbstractQName hierarchy has either a resolved QName, or two
unresolved {Qualified,Unqualified}QNames. The quality of being resolved
vs. unresolved is not modeled.

Introduce UnresolvedQName as the only AbstractQName alternative to
QName, so that it can be used to model this distinction. This allows,
for example, yang-xpath-api to be specified better.

JIRA: YANGTOOLS-1329
Change-Id: I9893850cea27bdd98d08b00ceb8c1f288dcc2778
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>