WIP: Populate ietf-restconf operations container
The 'operations' container defined in ietf-restconf module has magic
semantics: it needs to have an empty leaf for every RPC that is
available in the EffectiveModelContext.

WWe are hard-pressed here to make things work with our parser
infrastructure, as we currently do not have hooks to do this kind of
schema-tree specific magic. Luckily there is also a YANG extension in
that same model, so we hijack that to hook onto to the container
inference and modify it just in time and make the magic happen.

// FIXME: add typedef support for undeclared statements
// FIXME: finish test suite

JIRA: YANGTOOLS-1338
Change-Id: I9cf774a148f0e764940b9725f717fa0c60ef1bf9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>