WIP: Create leaf with empty value instead of null
There are some circumstances when we do not have
any other child, for example when we issue a delete
request from RESTCONF to delete a leaf. It means
we have no value (null) for the leaf builder.

But leaf nodes cannot accept null values.
Instead they have to be created with "empty" value.

Add "empty" value to leaf builder in this case to
successfully build an empty leaf.

JIRA: YANGTOOLS-1367
Change-Id: I0743464397d2c712048bb605a64f080605513f7a
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>