Add StatementSupportNamespace
Expose StatementSupports registered for a particular root from a new
namespace, StatementSupportNamespace. SourceSpecificContext already has
all the knowledge and wiring required to make this work, so it is a
rather straightforward exercise.

This immediately allows us to loosen couplig betwen action/rpc and
input/output supports. Unfortunately choice/case coupling still remains
as we do not get a reference to the statement context in execution path.

JIRA: YANGTOOLS-1371
Change-Id: Iccf8fd394c5bf172cda73c820f1a483b04bbc7a4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>