Hide StatementContextBase
We have an ugly dependency on StatementContextBase in statement
implementations. These should not care about implementation details,
and in fact they do not, as most of the methods involved are already
present in StmtContext.Mutable.

Expose the rest of the required methods and hide the base class. This
has a nice side effect of removing a significant number of unsafe casts.

JIRA: YANGTOOLS-1382
Change-Id: Id44e39dbedcc72089aa276f54ddffee4c60bca4d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>