Move StatementContextWriter.exitStatement()
The logic for entering a declared statement lives in
AbstractResumedStatement, along with the logic to create implicit
parents (i.e. implicit case statements). Also move the code to exit
the declared statement there, so we have all three bits in one place.

Since we have better context for dealing with this, also reshape the
logic, which allows us to move StatementContextBase.endDeclared().

JIRA: YANGTOOLS-1383
Change-Id: Id461464587b997f26624271f9bdf6aa9967f80a9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>