WIP: Integrate EffectiveAugmentationSchema
EffectiveAugmentationSchema is an oft-forgotten necessity when dealing
with AugmentationSchemaNode. It provides a bridge between
AugmentationSchemaNode and its manifestation at augmentation target
node.

The problem with it is that AugmentationSchemaNode and
AugmentEffectiveStatement do not hint at the critical difference between
presented substatements -- hence downstreams users are often faced with
having to fish for what is going on and the additional utility they have
to use to correctly interpret modeled data.

This patch alleviates that problem by properly documenting the diffences
in the view presented by manifestations of 'augment' statement and
provides a built-in means to seamlessly transition between the two
views.

// FIXME: finish integration

JIRA: YANGTOOLS-1403
Change-Id: Ief6f55726f1e0a68bf11824caf2a3c05783d59f8
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>