Invert SchemaInferenceStack.deque
While the use of the deque as a stack is kind of easy, it results in us
needing to operate on descendingIterator() rather than on Colleciton.
Invert the organization of the deque so that its iteration order
reflects the path from root. This has a number of benefits in terms of
code density.

JIRA: YANGTOOLS-1422
Change-Id: I442ef43cfe52842e334b48d32b901e5266041f59
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>