WIP: Integrate IncrementalBuildSupport
IncrementalBuildSupport guides incremental/resumed execution, in that it
compares inputs to previous state as well as carefully updating output
to match the intended codegen output.

The process is sensitive to losing previously-generated files, but
allows persistent files to be updated. If some files magically
appear/disappear, we trigger a generation cycle.

// FIXME: remove RebuildContext
// FIXME: update/rename RebuildContextTest
// FIXME: careful output file updates: do not rewrite files which are
//        already up-to-date, which requires collecting their bodies
//        first and then outputting them in one go -- which needs a
//        further refactor and parallel thread-safe update via
//        GeneratorTaks

JIRA: YANGTOOLS-745
Change-Id: I6f8cbf03a52542e8c20fd1383d459056272f951a
Signed-off-by: Ruslan Kashapov <ruslan.kashapov@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>